ci-test-quality wave 5: split AnalysisPanel and OverviewTab test files by logical group
